[Php74][TypeDeclaration] Add InlineVarDocToTypedPropertyRector + InlineVarDocToTypedInPrivatePropertyRector, deprecate TypedPropertyRector
