+++ rules/TypeDeclaration/Rector/Property/InlineVarDocToTypedInPrivatePropertyRector.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\TypeDeclaration\Rector\Property;
+
+use Rector\TypeDeclaration\Rector\AbstractInlineVarDocToTypedPropertyRector;
+
+final class InlineVarDocToTypedInPrivatePropertyRector extends AbstractInlineVarDocToTypedPropertyRector
+{
+}
